Add setup flow termination event
