+++ common/TelemetryEvents/SetupFlow/AppInstallEvent.cs
@@ -1,0 +1,31 @@
+﻿// Copyright (c) Microsoft Corporation and Contributors
+// Licensed under the MIT license.
+
+using System;
+using System.Diagnostics.Tracing;
+using DevHome.Telemetry;
+using Microsoft.Diagnostics.Telemetry;
+using Microsoft.Diagnostics.Telemetry.Internal;
+
+namespace DevHome.Common.TelemetryEvents.SetupFlow;
+
+[EventData]
+public class AppInstallEvent : EventBase
+{
+    public string PackageId { get; }
+
+    public string SourceId { get; }
+
+    public override PartA_PrivTags PartA_PrivTags => PrivTags.ProductAndServiceUsage;
+
+    public AppInstallEvent(string packageId, string sourceId)
+    {
+        PackageId = packageId;
+        SourceId = sourceId;
+    }
+
+    public override void ReplaceSensitiveStrings(Func<string, string> replaceSensitiveStrings)
+    {
+        // No sensitive strings to replace
+    }
+}
