+++ common/TelemetryEvents/SetupFlow/NavigateFromSummaryEvent.cs
@@ -7,22 +7,25 @@ using DevHome.Telemetry;
 using Microsoft.Diagnostics.Telemetry;
 using Microsoft.Diagnostics.Telemetry.Internal;
 
-namespace DevHome.Common.TelemetryEvents;
+namespace DevHome.Common.TelemetryEvents.SetupFlow;
 
 [EventData]
-public class LoadingRetryEvent : EventBase
+public class NavigateFromSummaryEvent : EventBase
 {
-    public int NumberOfFailedTasks { get; }
+    public string NavigationTarget
+    {
+        get;
+    }
 
     public override PartA_PrivTags PartA_PrivTags => PrivTags.ProductAndServiceUsage;
 
-    public LoadingRetryEvent(int numberOfFailedTasks)
+    public NavigateFromSummaryEvent(string navigationTarget)
     {
-        NumberOfFailedTasks = numberOfFailedTasks;
+        NavigationTarget = navigationTarget;
     }
 
     public override void ReplaceSensitiveStrings(Func<string, string> replaceSensitiveStrings)
     {
-        // No sensitive data held
+        // No sensitive strings to replace
     }
 }
