+++ tools/SetupFlow/DevHome.SetupFlow/Views/RepoConfigView.xaml.cs
@@ -5,13 +5,9 @@ using System;
 using System.Linq;
 using DevHome.Common.Extensions;
 using DevHome.Common.Models;
-using DevHome.Common.TelemetryEvents;
-using DevHome.Common.TelemetryEvents.RepoToolEvents;
-using DevHome.Common.TelemetryEvents.RepoToolEvents.EditDialogEvents;
-using DevHome.Common.TelemetryEvents.RepoToolEvents.RepoDialog;
+using DevHome.Common.TelemetryEvents.SetupFlow;
 using DevHome.Contracts.Services;
 using DevHome.SetupFlow.Models;
-using DevHome.SetupFlow.Services;
 using DevHome.SetupFlow.ViewModels;
 using DevHome.Telemetry;
 using Microsoft.UI.Xaml;
